Extract websocket events to the separate controller
