+++ controllers/web/draft_controller.js
@@ -22,7 +22,7 @@ router.get("/draft/:draftId/:side", async (req, res) => {
 			blueTeamName: draft.blueTeamName,
 			redTeamName: draft.redTeamName,
 			pickTimeout: draft.options?.pickTimeout || draft.pickTimeout,
-			nicknames: draft.options?.nicknames  || draft.nicknames,
+			nicknames: draft.options?.nicknames || draft.nicknames,
 		});
 	} catch (error) {
 		console.log(`Error rendering draft: ${error.message}`);
